Move GalleryGroup to TS, add incremental loading
